move modo to single module
