+++ modo-render-android-compose/src/main/java/com/github/terrakok/modo/ComposeScreen.kt
@@ -1,0 +1,17 @@
+package com.github.terrakok.modo
+
+import android.os.Parcelable
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.staticCompositionLocalOf
+import com.github.terrakok.modo.containers.ContainerScreen
+
+interface Screen : Parcelable {
+
+    @Composable
+    fun Content()
+
+    val screenKey: String
+
+    val screenName: String get() = "Screen"
+
+}
