+++ modo-render-android-compose/src/main/java/com/github/terrakok/modo/Modo.kt
@@ -1,4 +1,4 @@
-package com.github.terrakok.modo.android.compose
+package com.github.terrakok.modo
 
 import android.os.Bundle
 
@@ -7,7 +7,7 @@ object Modo {
     private const val MODO_SCREEN_COUNTER_KEY = "MODO_SCREEN_COUNTER_KEY"
     private const val MODO_GRAPH = "MODO_GRAPH"
 
-    fun saveInstanceState(outState: Bundle, rootScreen: ComposeScreen) {
+    fun saveInstanceState(outState: Bundle, rootScreen: Screen) {
         outState.putInt(MODO_SCREEN_COUNTER_KEY, screenCounterKey.get())
         outState.putParcelable(MODO_GRAPH, rootScreen)
     }
@@ -16,10 +16,10 @@ object Modo {
      * @param savedState
      * @param rootScreenProvider invokes whe saved state is null and we need to provide root screen
      */
-    fun <T: ComposeScreen> restoreInstanceIfCan(savedState: Bundle?, rootScreenProvider: () -> T): T =
+    fun <T: Screen> restoreInstanceIfCan(savedState: Bundle?, rootScreenProvider: () -> T): T =
         if (savedState != null) {
             restoreScreenCounter(savedState.getInt(MODO_SCREEN_COUNTER_KEY))
-            savedState.getParcelable<ComposeScreen>(MODO_GRAPH) as T
+            savedState.getParcelable<Screen>(MODO_GRAPH) as T
         } else {
             rootScreenProvider()
         }
