+++ sample/src/main/java/com/github/terrakok/androidcomposeapp/CustomAction.kt
@@ -1,12 +1,11 @@
 package com.github.terrakok.androidcomposeapp
 
 import com.github.terrakok.modo.*
-import com.github.terrakok.modo.android.compose.ComposeScreen
-import com.github.terrakok.modo.android.compose.Stack
+import com.github.terrakok.modo.Screen
 
 class AddTab(
     val id: String,
-    val rootScreen: ComposeScreen
+    val rootScreen: Screen
 ) : NavigationAction
 
 class CustomReducer : NavigationReducer<MultiNavigation> {
